refactor: 하이라이트 훅 폴더 위치 변경
- HighlightEditor 아래로 위치 변경
