+++ frontend/src/components/highlight/components/HighlightEditor/hooks/useCheckHighlight.ts
