+++ frontend/src/components/highlight/components/HighlightEditor/hooks/useDragHighlightButtonPosition.ts
